trigger/requirements.txt
fastapi==0.136.0
uvicorn[standard]==0.45.0
livekit-api==1.1.0
python-dotenv==1.2.2
